+++ controllers/GlobalController.php
@@ -4,7 +4,11 @@ namespace humhub\modules\calendar\controllers;
 
 use DateTime;
 use humhub\components\Controller;
+use humhub\modules\calendar\helpers\CalendarUtils;
 use humhub\modules\calendar\interfaces\CalendarService;
+use humhub\modules\calendar\interfaces\recurrence\RecurrenceFormModel;
+use humhub\modules\calendar\models\CalendarEntry;
+use humhub\modules\calendar\models\CalendarEntryDummy;
 use humhub\modules\calendar\models\fullcalendar\FullCalendar;
 use humhub\modules\calendar\models\SnippetModuleSettings;
 use humhub\modules\calendar\permissions\CreateEntry;
@@ -269,5 +273,12 @@ class GlobalController extends Controller
             'footer' => $enableButton . $nextButton . $cancelButton,
             'centerText' => true
         ]);
+    }
+
+    public function actionUpdateMonthlyRecurrenceSelection($date)
+    {
+        $dummy = new CalendarEntryDummy(['start' => CalendarUtils::parseDateTimeString($date, null, null, null)]);
+        $recurrenceForm = new RecurrenceFormModel(['entry' => $dummy]);
+        return $this->asJson(['result' => $recurrenceForm->getMonthDaySelection()]);
     }
 }